+++ wayland-backend/src/rs/server/client.rs
@@ -172,14 +172,14 @@ impl<D> Client<D> {
                     Argument::Object(o.id)
                 },
                 Argument::Object(o) => {
+                    let next_interface = arg_interfaces.next().unwrap();
                     if o.id != 0 {
                         if o.client_id != self.id {
                             panic!("Attempting to send an event with objects from wrong client.")
                         }
-                        let object = self.get_object(o.clone())?;
-                        let next_interface = arg_interfaces.next().unwrap();
-                        if !same_interface_or_anonymous(next_interface, object.interface) {
-                            panic!("Event {}@{}.{} expects an object argument of interface {} but {} was provided instead.", object.interface.name, object_id.id, message_desc.name, next_interface.name, object.interface.name);
+                        let arg_object = self.get_object(o.clone())?;
+                        if !same_interface_or_anonymous(next_interface, arg_object.interface) {
+                            panic!("Event {}@{}.{} expects an object argument of interface {} but {} was provided instead.", object.interface.name, object_id.id, message_desc.name, next_interface.name, arg_object.interface.name);
                         }
                     } else if !matches!(message_desc.signature[i], ArgumentType::Object(AllowNull::Yes)) {
                             panic!("Request {}@{}.{} expects an non-null object argument.", object.interface.name, object_id.id, message_desc.name);